+++ functions/src/articles.js
@@ -5,7 +5,8 @@ const {
   handleResponse, 
   handleError, 
   validateRequired, 
-  sanitizeString
+  sanitizeString,
+  verifyAdminAuth
 } = require("./utils/helpers");
 const { uploadImageFromBase64, deleteImage } = require("./services/imageService");
 
@@ -211,14 +212,30 @@ const articles = onRequest(async (req, res) => {
           return await getArticleById(pathParts[0], req, res);
         }
       } else if (method === 'POST') {
+        // Require admin auth for all POST operations
+        const isAdmin = await verifyAdminAuth(req);
+        if (!isAdmin) {
+          return handleResponse(res, { message: 'Admin access required' }, 403);
+        }
+        
         if (pathParts.length === 0) {
           return await createArticle(req, res);
         } else if (pathParts.length === 2 && pathParts[1] === 'upload-image') {
           return await uploadArticleImage(pathParts[0], req, res);
         }
       } else if (method === 'PUT' && pathParts.length === 1) {
+        // PUT /articles/:id - Require admin auth
+        const isAdmin = await verifyAdminAuth(req);
+        if (!isAdmin) {
+          return handleResponse(res, { message: 'Admin access required' }, 403);
+        }
         return await updateArticle(pathParts[0], req, res);
       } else if (method === 'DELETE' && pathParts.length === 1) {
+        // DELETE /articles/:id - Require admin auth
+        const isAdmin = await verifyAdminAuth(req);
+        if (!isAdmin) {
+          return handleResponse(res, { message: 'Admin access required' }, 403);
+        }
         return await deleteArticle(pathParts[0], req, res);
       }
